+++ protocols/gossipsub/CHANGELOG.md
@@ -15,6 +15,7 @@
   [PR 2103](https://github.com/libp2p/rust-libp2p/pull/2103)
 
 - Added std::error::Error impls for error types
+  [PR 2254](https://github.com/libp2p/rust-libp2p/pull/2254)
 
 # 0.32.0 [2021-07-12]