+++ protocols/gossipsub/CHANGELOG.md
@@ -14,7 +14,7 @@
 - Allow `message_id_fn`s to accept closures that capture variables.
   [PR 2103](https://github.com/libp2p/rust-libp2p/pull/2103)
 
-- Added std::error::Error impls for error types
+- Implement std::error::Error for error types.
   [PR 2254](https://github.com/libp2p/rust-libp2p/pull/2254)
 
 # 0.32.0 [2021-07-12]