+++ protocols/gossipsub/CHANGELOG.md
@@ -14,6 +14,9 @@
 - Allow `message_id_fn`s to accept closures that capture variables.
   [PR 2103](https://github.com/libp2p/rust-libp2p/pull/2103)
 
+- Implement std::error::Error for error types.
+  [PR 2254](https://github.com/libp2p/rust-libp2p/pull/2254)
+
 # 0.32.0 [2021-07-12]
 
 - Update dependencies.