+++ protocols/gossipsub/CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.33.0 [unreleased]
+# 0.33.0-rc.1 [2021-10-15]
 
 - Add an event to register peers that do not support the gossipsub protocol
   [PR 2241](https://github.com/libp2p/rust-libp2p/pull/2241)